// packages/stellar/src/generated/settlement.ts
export interface PaymentIntent {
  amount: i128;
  customer: string;
  expires_at_ledger: u32;
  intent_id: Buffer;
  merchant_id: Buffer;
  network_id: Buffer;
  nonce: Buffer;
  recipient: string;
  settlement_contract: string;
  token: string;
}